Query Middleware: Clone expressions more efficiently (#12969)
#### What this PR does

In many places Query Middleware copies the data structure for a PromQL
expression, so that it can be modified in private.
Instead of doing this copy by rendering to a string then parsing the
string, use the `CloneExpr` function introduced in #12787.

 - Export `astmapper.CloneExpr` (was `cloneExpr`) and update all
internal calls.
 - Add `MetricsQueryRequest.GetParsedQuery()`; implement in
`PrometheusRangeQueryRequest`, `PrometheusInstantQueryRequest`, and
remote-read request.
 - Prefer `WithExpr(parser.Expr)` over `WithQuery(string)` where
rewriting/mapping occurs.