distributor, ruler: rename "organization" span attribute to be consistent with logs (#13487)
#### What this PR does

The PR updates two instances, where we add the `organization=<userID>`
attributes to the spans. Mimir tends to use the term `user` across the
code base, when the `userID` is logged. I don't see why we should be
doing it differently in the traces.

Note (1): I looked across the git history, and both instances, the PR
touches, were added 4-5 years ago.

Note (2): This is a minor change, that I don't want to bother adding to
the changelog.

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>