storegateway: add cortex_bucket_store_block_discovery_latency_seconds (#13489)
#### What this PR does

This PR add an experimental
`cortex_bucket_store_block_discovery_latency_seconds` metric, that
tracks how much time it took for the store-gateway to discover a block,
from the time of the block's creation.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>