Fix issue where utilization-based limiter always fails to start (#13541)
#### What this PR does

This PR fixes an issue introduced in
https://github.com/grafana/mimir/pull/13408 where the utilization-based
limiter always fails to start with `unable to detect CPU/memory
utilization, unsupported platform. Please disable utilization based
limiting: %!w(<nil>)`

I've chosen not to add a changelog entry given #13408 does not have a
changelog entry and this change has not been part of a numbered release
yet.

#### Which issue(s) this PR fixes or relates to

#13408

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.