MQE: add metrics for the narrow binary selectors optimization pass (#13533)
Add metrics for when we attempt to add hints to binary expressions and
when we are actually able to add hints to binary expressions. Queries
that have been rewritten by query-frontend middleware are not included
in the count of "attempts" to add hints since we know that we can't
modify them.