MQE: Skip using labels for binary expression hints based on children
Change the logic introduced in #13499 to only skip using particular
labels for hints in binary expressions when the label is created by
a function call that is a child of that binary expression. This allows
us to apply hints in more cases when it is safe to do so.

Related #13499

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>