usage-tracker: refactor/rename some tenantshard implementation details
Follow up on https://github.com/grafana/mimir/pull/13594

I'm removing the word 'meta' because we don't call it that way in our implementation, we call it index. And H2 is 'prefix' in our impl, and if we make this method a receiver, we don't need to be so verbose in the name because the signature of index.match(prefix) already tells what it's doing.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>