MQE: Skip using labels for binary expression hints based on children (#13592)
Change the logic introduced in #13499 to only skip using particular
labels for hints in binary expressions when the label is created by a
function call that is a child of that binary expression. This allows us
to apply hints in more cases when it is safe to do so.

Related #13499

Suggested in
https://github.com/grafana/mimir/pull/13499#discussion_r2536013342