index planning: fix metrics
i had missed fixing up the type assertions in XXX so now we aren't getting metrics for planning outcomes

this PR just removes type assertions. It also removes a case where there isn't an error but the returned plan is nil. The only case where this could have happened before is when there was a panic. In this case with this PR we would just record some bogus metrics.