helm: default image tag from appVersion (#13453)
#### What this PR does

As originally suggested in
https://github.com/grafana/mimir/pull/13291/commits/97c4ae81c8604e47b1ed15be3598b58461ab4638
the PR updates the Helm chart to default the container image's tag to
the value of chart's AppVersion. This will make managing the future
releases simpler, because the release manager won't need to bother
updating the values.

Note, in the follow-up PR I'm updating the CI job that updates the
weekly releases (ref. to [the
comment](https://github.com/grafana/mimir/pull/13291#issuecomment-3480866237))
— see https://github.com/grafana/mimir/pull/13729

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
Co-authored-by: Niklas Wagner <Skaro@Skaronator.com>