Align MQE histogram counter conflict logic to prometheus (#13623)
#### What this PR does

Do not merge until [upstream
vendoring](https://github.com/prometheus/prometheus/pull/17608) is
vendored in.

This PR aligns the histogram counter reset conflict detection logic to
match the prometheus implementation for avg and sum aggregations. It
also fixes an issue where the histogram counter reset hint was not set
to Guage on SUB operations.

This PR also adds a unit test to `querysharding_test.go`
`TestConflictingCounterResets` which validates a fix for an issue with
query-frontends sharding queries to prometheus promql engine.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [ x] Tests updated.
- [ ] Documentation added.
- [ x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> <sup>[Cursor Bugbot](https://cursor.com/dashboard?tab=bugbot) is
generating a summary for commit
6114d0f248b6d5324431d573e0f3e60318182122. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->