Update MQE team members tagged when Prometheus vendoring PR fails `pkg/streamingpromql` tests (#13786)
#### What this PR does

This PR updates the list of people tagged in a comment like [this
one](https://github.com/grafana/mimir/pull/13784#issuecomment-3629128759)
when a Prometheus vendoring PR fails unit tests in the
`pkg/streamingpromql` package.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.