compactor: verify that TSDB uploads use index format v2 (#13815)
Verify that the TSDB block index file uses v2 for the file format, not
v1. The v1 format was only used for Prometheus 2.0 and 2.1 as far as I
can tell. It hasn't been the default format for new blocks since
Prometheus 2.2 and has never been the default for Mimir.

Part of #13808