store-gateway: remove support for reading TSDB index v1
This change removes support for generating index-headers and reading
the v1 format of TSDB index files. The v1 format was only the default
in Prometheus 2.0 and 2.1. Prometheus 2.2 and later default to creating
blocks that use the v2 format. The v1 format has never been the default
in Mimir.

Note that this change does not refactor or change index-header code
beyond removing v1 support. Further cleanups will be done in follow
up PRs.

Part of #13808

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>