CHANGELOG: Downgrade some CHANGEs to ENHANCEMENTS
Where a change was purely updating a feature from experimental to stable
(or "advanced"), don't list it as a CHANGE.

CHANGE is supposed to be for breaking changes, that an admin must know
about before upgrading.

The Go 1.25.4 update is removed entirely, since it did not contain any
CVE fixes and there is already a line for Go 1.25.5.