store-gateway: Add a test to simulate the RF3 -> RF4 migration
When migrating from replication factor 3 to 4 in a zone-aware deployment, it's critical that blocks are not reshuffled in the zones we don't touch. Block reshuffling causes:

1. Query failures when queriers hit store-gateways that don't have the expected blocks
2. Unnecessary data transfer and disk I/O as store-gateways download index-headers for reshuffled block

The test verifies that throughout this process, blocks owned by zone-a and zone-b instances never change. Zone-c is added and re-add because it simulates what happens when we move zone-c to a different AZ.

Signed-off-by: gotjosh <josue.abreu@gmail.com>