MQE: Don't apply narrow selectors pass to logical or/unless (#13837)
Don't set hints to apply extra selectors to the RHS of a binary
operation based on results from the LHS if the binary operation is
`logical or` or `logical unless`. This isn't a problem currently because
those operations are handled by different operators that haven't yet
been changed to apply extra matchers.