ingester: add close-idle-tsdb-when-shipping-disabled (#13862)
#### What this PR does

This PR adds a new `close-idle-tsdb-when-shipping-disabled`
configuration flag to ingesters. When the shipment of blocks is
disabled, the ingesters continue accumulation idle users with 0 series.

This becomes problematic, when Mimir runs with an (experimental)
block-builder component, that takes the responsibility to build and ship
blocks to a long-term object storage. In this mode, the ingesters act as
a persistent cache for recent data. The ingesters don't need to ship
blocks, but they should continue to close idle tenants.

#### Which issue(s) this PR fixes or relates to

Relates to https://github.com/grafana/mimir-squad/issues/3410

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>